requirements.txt
discord.py>=2.0.0
fastf1>=2.3.0
matplotlib>=3.5.0
numpy>=1.20.0
pandas>=1.3.0
seaborn>=0.11.0
pytz>=2021.3
beautifulsoup4>=4.10.0
requests>=2.27.0
Pillow>=9.0.0
